Remove job context manager on SpawningWorker
The issue with this context manager is that is on `__exit__` it marks the job either complete or sets an error. But in this case, the error happens inside a child process and isn't propagated to the calling process. So essentially, even if the subprocess set this job as failed, it will be overridden to be completed.